# Conflicts:
#	src/Components/Agenda.js
#	src/Services/AgendaService.js
#	src/Services/LesService.js
#	src/Services/OptredenService.js

+++ src/Components/Instrumenten.js
@@ -5,58 +5,81 @@ import React, {Component} from 'react';
 import * as InstrumentenService from '../Services/InstrumentService.js'
 import {Link} from 'react-router-dom';
 import Header from './Header'
-import * as LoginService from "../Services/LoginService";
-import Redirect from "react-router-dom/es/Redirect";
+import swal from 'sweetalert2'
+
+
 class Instrumenten extends Component {
 
     constructor(props) {
         super(props);
+        console.log("Constructed");
         this.state = {
             instrumenten: [],
             selectedIndex: 0,
-            selected: [],
         };
     }
 
     handleDelete = (id, e) => {
-        InstrumentenService.deleteInstrument(id);
+        swal({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            type: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Delete!',
+            cancelButtonText: 'Cancel!',
+            confirmButtonClass: 'btn red',
+            cancelButtonClass: 'btn green marginator',
+            buttonsStyling: false,
+            reverseButtons: true
+        }).then((result) => {
+            if (result.value) {
+                swal(
+                    'Deleted!',
+                    'Instrument has been deleted.',
+                    'success'
+                );
+                InstrumentenService.deleteInstrument(id);
+            } else if (
+                // Read more about handling dismissals
+            result.dismiss === swal.DismissReason.cancel
+            ) {
+                swal(
+                    'Cancelled',
+                    'Instrument was not deleted',
+                    'error'
+                )
+            }
+        });
+
     };
 
-
-    componentWillReceiveProps() {
+    getInstrumenten() {
         InstrumentenService.getInstrumentenFromBackend().then(instrumenten => {
             this.setState({instrumenten: instrumenten});
         });
+    }
+
+
+    componentWillUpdate() {
+        this.getInstrumenten();
     }
 
 
     componentDidMount() {
-        InstrumentenService.getInstrumentenFromBackend().then(instrumenten => {
-            this.setState({instrumenten: instrumenten});
-        });
-    }
-
-    componentWillMount(){
-        let response = false;
-        response = LoginService.checkToken();
-        console.log("response:");
-        console.log(response);
-        this.setState({redirect: !response})
+       this.getInstrumenten();
     }
 
 
     render() {
-        let redirecter=null;
-        if (this.state.redirect) {
-            redirecter = <Redirect to='/login'/>
-        }
+
         return (
             <div className="Homepage">
-                {redirecter}
                 <Header name="Instrumenten" />
 
                 <section className="containerCss">
-                    <table className="white-text bordered responsive-table centered">
+                    <table className="highlight striped black-text bordered responsive-table centered">
                         <thead>
                         <tr>
                             <th>Id</th>
@@ -86,7 +109,11 @@ class Instrumenten extends Component {
                         ))}
                         </tbody>
                     </table>
-
+                    <div className="fixed-action-btn">
+                        <Link to="/addInstrument" className="btn-floating btn-large deep-orange darken-4">
+                            <i className="large material-icons">add</i>
+                        </Link>
+                    </div>
                 </section>
             </div>
         );
